requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis==5.0.4
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
alembic==1.13.1
httpx==0.27.0
python-multipart==0.0.9
motor==3.4.0
pymongo==4.6.3
bcrypt==3.2.2